requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.7.1
ecdsa==0.17.0
entrypoints==0.3
fastapi==0.68.0
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1

ipywidgets==7.6.3

Jinja2==3.0.1
joblib==1.0.1
jose==1.0.0
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1

mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy~=1.19.5
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3

passlib==1.7.4

Pillow==8.3.1
pluggy==0.13.1
prometheus-client==0.11.0

protobuf==3.17.3

py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2

pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
pyzmq==20.0.0
requests==2.26.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0

sklearn==0.0
smmap==4.0.0

SQLAlchemy==1.4.23
starlette==0.14.2
streamlit==0.87.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1

typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
uvicorn==0.15.0
validators==0.18.2
watchdog==2.1.5

webencodings==0.5.1
widgetsnbextension==3.5.1

autopep8==1.5.7
greenlet==1.1.1
mysql-connector-python==8.0.26
numpy==1.21.2
protobuf==3.17.3
pycodestyle==2.7.0
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2

keras~=2.6.0